requirements.txt
Django==5.2.11

django-environ==0.12.0
djangorestframework==3.15.2
psycopg[binary]==3.2.6
openpyxl==3.1.5
google-api-python-client==2.168.0
google-auth==2.40.1
